+++ src/aks-preview/azext_aks_preview/commands.py
@@ -16,6 +16,7 @@ from ._format import aks_versions_table_format
 from ._format import aks_upgrades_table_format
 from ._format import aks_pod_identities_table_format
 from ._format import aks_pod_identity_exceptions_table_format
+from ._format import aks_run_command_result_format
 
 
 def load_command_table(self, _):
@@ -52,22 +53,34 @@ def load_command_table(self, _):
         g.custom_command('create', 'aks_create', supports_no_wait=True)
         g.custom_command('update', 'aks_update', supports_no_wait=True)
         g.custom_command('scale', 'aks_scale', supports_no_wait=True)
-        g.custom_command('disable-addons', 'aks_disable_addons', supports_no_wait=True)
-        g.custom_command('enable-addons', 'aks_enable_addons', supports_no_wait=True)
+        g.custom_command('disable-addons', 'aks_disable_addons',
+                         supports_no_wait=True)
+        g.custom_command('enable-addons', 'aks_enable_addons',
+                         supports_no_wait=True)
         g.custom_command('get-credentials', 'aks_get_credentials')
-        g.custom_show_command('show', 'aks_show', table_transformer=aks_show_table_format)
+        g.custom_show_command('show', 'aks_show',
+                              table_transformer=aks_show_table_format)
         g.custom_command('upgrade', 'aks_upgrade', supports_no_wait=True)
-        g.command('get-upgrades', 'get_upgrade_profile', table_transformer=aks_upgrades_table_format)
+        g.command('get-upgrades', 'get_upgrade_profile',
+                  table_transformer=aks_upgrades_table_format)
         g.custom_command('rotate-certs', 'aks_rotate_certs', supports_no_wait=True,
                          confirmation='Kubernetes will be unavailable during certificate rotation process.\n' +
                          'Are you sure you want to perform this operation?')
         g.wait_command('wait')
         g.command('stop', 'stop', supports_no_wait=True)
         g.command('start', 'start', supports_no_wait=True)
+        g.custom_command('get-os-options', 'aks_get_os_options')
 
     # AKS container service commands
     with self.command_group('aks', container_services_sdk, client_factory=cf_container_services) as g:
-        g.custom_command('get-versions', 'aks_get_versions', table_transformer=aks_versions_table_format)
+        g.custom_command('get-versions', 'aks_get_versions',
+                         table_transformer=aks_versions_table_format)
+
+    with self.command_group('aks command', managed_clusters_sdk, client_factory=cf_managed_clusters) as g:
+        g.custom_command('invoke', 'aks_runcommand', supports_no_wait=True,
+                         table_transformer=aks_run_command_result_format)
+        g.custom_command('result', 'aks_command_result',
+                         supports_no_wait=False, table_transformer=aks_run_command_result_format)
 
     # AKS maintenance configuration commands
     with self.command_group('aks maintenanceconfiguration', maintenance_configuration_sdk, client_factory=cf_maintenance_configurations) as g:
@@ -79,13 +92,18 @@ def load_command_table(self, _):
 
     # AKS agent pool commands
     with self.command_group('aks nodepool', agent_pools_sdk, client_factory=cf_agent_pools) as g:
-        g.custom_command('list', 'aks_agentpool_list', table_transformer=aks_agentpool_list_table_format)
-        g.custom_show_command('show', 'aks_agentpool_show', table_transformer=aks_agentpool_show_table_format)
+        g.custom_command('list', 'aks_agentpool_list',
+                         table_transformer=aks_agentpool_list_table_format)
+        g.custom_show_command('show', 'aks_agentpool_show',
+                              table_transformer=aks_agentpool_show_table_format)
         g.custom_command('add', 'aks_agentpool_add', supports_no_wait=True)
         g.custom_command('scale', 'aks_agentpool_scale', supports_no_wait=True)
-        g.custom_command('upgrade', 'aks_agentpool_upgrade', supports_no_wait=True)
-        g.custom_command('update', 'aks_agentpool_update', supports_no_wait=True)
-        g.custom_command('delete', 'aks_agentpool_delete', supports_no_wait=True)
+        g.custom_command('upgrade', 'aks_agentpool_upgrade',
+                         supports_no_wait=True)
+        g.custom_command('update', 'aks_agentpool_update',
+                         supports_no_wait=True)
+        g.custom_command('delete', 'aks_agentpool_delete',
+                         supports_no_wait=True)
         g.custom_command('get-upgrades', 'aks_agentpool_get_upgrade_profile')
 
     # AKS pod identity commands