+++ src/datadog/azext_datadog/vendored_sdks/datadog/_configuration.py
@@ -28,7 +28,7 @@ class MicrosoftDatadogClientConfiguration(Configuration):
 
     :param credential: Credential needed for the client to connect to Azure.
     :type credential: ~azure.core.credentials.TokenCredential
-    :param subscription_id: The Microsoft Azure subscription ID.
+    :param subscription_id: The ID of the target subscription.
     :type subscription_id: str
     """
 
@@ -47,9 +47,8 @@ class MicrosoftDatadogClientConfiguration(Configuration):
 
         self.credential = credential
         self.subscription_id = subscription_id
-        self.api_version = "2020-02-01-preview"
-        self.credential_scopes = ['https://management.azure.com/.default']
-        self.credential_scopes.extend(kwargs.pop('credential_scopes', []))
+        self.api_version = "2021-03-01"
+        self.credential_scopes = kwargs.pop('credential_scopes', ['https://management.azure.com/.default'])
         kwargs.setdefault('sdk_moniker', 'microsoftdatadogclient/{}'.format(VERSION))
         self._configure(**kwargs)